Created queue functionality. Now, the sessions can be grouped together and use the same queue to pool workers.
The queue assigns priorities to the tickets in the queue and the worker will receive the items in the increasing order.

TODO: from time to time we should reset the worker queue positions and make them plow through the still-active tasks that they haven't done. Or the new tasks with higher priority.